requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==25.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
Events==0.5
fastapi==0.117.1
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
h11==0.16.0
haystack-ai==2.18.0
haystack-experimental==0.13.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-imports==0.3.1
MarkupSafe==3.0.3
more-itertools==10.8.0
neo4j==5.28.2
neo4j-haystack==2.2.1
networkx==3.5
num2words==0.5.14
numpy==2.3.3
ollama==0.6.0
ollama-haystack==5.2.0
openai==1.109.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
posthog==6.7.6
prompthub-py==4.0.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-cache==0.9.8
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.56.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
